Temporary fix for the logo being affected by the background
The problem with this is that the texture used for the logo is incorrect. The fully transparent pixels have a black color, causing artifacts along the edge when the logo is scaled. The only way to fix this is to make the logo fully white (preserving the alpha channel), but I'm not quite sure how to do that at the moment. As such this fix is only half complete.