Fix build
In order for this to work on a real client, the Redirect needs to be remapped, but, uh... apparently Mixin does not know the existence of the inner class LogoTexture, so I have to provide the full signature for it to look up in its mappings. With this change the jar builds and runs in a real client.