Clamp texture coordinates to edge
This prevents the bilinear scaling from creating small bars on either side of the 'STUDIOS' subtext of the logo.

This is because the logo is separated into two halves, and the split happens to be right in the middle of the left stroke of the 'D', causing that to be visible on both sides. It confused me for a while too. Then I happened to see an article on wrapping modes and I instantly saw what the problem was.